Create virtual env wrapper env
