+++ .github/workflows/pr-check.yml
@@ -95,6 +95,9 @@ jobs:
         run: |
           pip install virtualenvwrapper
           echo "WORKON_HOME=$HOME/.virtualenvs" >> $GITHUB_ENV
+          mkdir -p $WORKON_HOME
+          source /usr/local/bin/virtualenvwrapper.sh
+          mkvirtualenv venv_wrapper_env1
         shell: bash
 
       - name: Install virtualenvwrapper-win
