+++ .gitignore
@@ -8,3 +8,6 @@ target/
 
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
+
+# Directory with generated environments (generally created on CI)
+tmp
