+++ crates/pet-utils/src/executable.rs
@@ -99,30 +99,6 @@ fn is_python_executable_name(exe: &Path) -> bool {
     }
 }
 
-// Resolves symlinks to the real file.
-// If the real file == exe, then it is not a symlink.
-pub fn resolve_symlink(exe: &Path) -> Option<PathBuf> {
-    let name = exe.file_name()?.to_string_lossy();
-    // In bin directory of homebrew, we have files like python-build, python-config, python3-config
-    if !name.starts_with("python") || name.ends_with("-config") || name.ends_with("-build") {
-        return None;
-    }
-    if let Ok(metadata) = std::fs::symlink_metadata(exe) {
-        if metadata.is_file() || !metadata.file_type().is_symlink() {
-            return Some(exe.to_path_buf());
-        }
-        if let Ok(readlink) = std::fs::canonicalize(exe) {
-            if readlink == exe {
-                return None;
-            } else {
-                return Some(readlink);
-            }
-        }
-        return Some(exe.to_path_buf());
-    }
-    Some(exe.to_path_buf())
-}
-
 // Given a list of executables, return the one with the shortest path.
 // The shortest path is the most likely to be most user friendly.
 pub fn get_shortest_executable(exes: &Option<Vec<PathBuf>>) -> Option<PathBuf> {
